replaced int with float and returned true. good

+++ report_card_printer.py
@@ -9,4 +9,4 @@ age = 20
 print(age, type(age))   
 
 score = 80.5
-print(isinstance(score, int))
+print(isinstance(score, float))
